☑️ Adding tests for handler.copy method
With this commit we're verifying that based on the input we're receiving
the expected output.

It's not quite ideal as we're relying on some inner workings of the
derivative rodeo.  We'd love to see what was put in the bucket as well
as the queue, but that's not quite where we're at.